Add more stuff to CustomAttachmentHandler
